Share fade in logic with repeats
