+++ osu.Game.Rulesets.Osu/Objects/SliderTailCircle.cs
@@ -1,7 +1,6 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using osu.Framework.Bindables;
 using osu.Game.Rulesets.Judgements;
 using osu.Game.Rulesets.Objects;
 using osu.Game.Rulesets.Osu.Judgements;
@@ -13,18 +12,8 @@ namespace osu.Game.Rulesets.Osu.Objects
     /// Note that this should not be used for timing correctness.
     /// See <see cref="SliderEventType.LegacyLastTick"/> usage in <see cref="Slider"/> for more information.
     /// </summary>
-    public class SliderTailCircle : SliderCircle
+    public class SliderTailCircle : SliderEndCircle
     {
-        private readonly IBindable<int> pathVersion = new Bindable<int>();
-
-        public SliderTailCircle(Slider slider)
-        {
-            pathVersion.BindTo(slider.Path.Version);
-            pathVersion.BindValueChanged(_ => Position = slider.EndPosition);
-        }
-
-        protected override HitWindows CreateHitWindows() => HitWindows.Empty;
-
         public override Judgement CreateJudgement() => new SliderTailJudgement();
 
         public class SliderTailJudgement : OsuJudgement
